pyproject.toml
[project]
name = "graphql"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "annotated-doc==0.0.4",
    "annotated-types==0.7.0",
    "anyio==4.12.1",
    "cachecontrol==0.14.4",
    "certifi==2026.2.25",
    "cffi==2.0.0 ; platform_python_implementation != 'PyPy'",
    "charset-normalizer==3.4.6",
    "click==8.3.1",
    "colorama==0.4.6 ; sys_platform == 'win32'",
    "cross-web==0.4.1",
    "cryptography==46.0.5",
    "fastapi==0.135.1",
    "firebase-admin==7.2.0",
    "google-api-core==2.30.0",
    "google-auth==2.49.1",
    "google-cloud-core==2.5.0",
    "google-cloud-firestore==2.25.0 ; platform_python_implementation != 'PyPy'",
    "google-cloud-storage==3.9.0",
    "google-crc32c==1.8.0",
    "google-resumable-media==2.8.0",
    "googleapis-common-protos==1.73.0",
    "graphql-core==3.2.8",
    "grpcio==1.78.0",
    "grpcio-status==1.78.0",
    "h11==0.16.0",
    "h2==4.3.0",
    "hpack==4.1.0",
    "httpcore==1.0.9",
    "httpx==0.28.1",
    "hyperframe==6.1.0",
    "idna==3.11",
    "msgpack==1.1.2",
    "packaging==26.0",
    "proto-plus==1.27.1",
    "protobuf==6.33.5",
    "pyasn1==0.6.3",
    "pyasn1-modules==0.4.2",
    "pycparser==3.0 ; implementation_name != 'PyPy' and platform_python_implementation != 'PyPy'",
    "pydantic==2.12.5",
    "pydantic-core==2.41.5",
    "pyjwt==2.12.1",
    "python-dateutil==2.9.0.post0",
    "python-multipart==0.0.22",
    "requests==2.32.5",
    "six==1.17.0",
    "starlette==0.52.1",
    "strawberry-graphql==0.311.3",
    "typing-extensions==4.15.0",
    "typing-inspection==0.4.2",
    "urllib3==2.6.3",
    "uvicorn==0.42.0",
]
